move angular project to resource dir

+++ app/Domains/Governo/Validators/StoreGovernoValidator.php
@@ -9,12 +9,12 @@ class StoreGovernoValidator extends Validator
     public function rules()
     {
         return [
-            'gasto' => ['required', 'float'],
-            'receita' => ['required', 'float'],
-            'imposto_renda' => ['required', 'float'],
-            'taxa_juros' => ['required', 'float'],
-            'taxa_deposito_compulsorio' => ['required', 'float'],
-            'salario_minimo' => ['required', 'float'],
+            'gasto' => ['required', 'numeric'],
+            'receita' => ['required', 'numeric'],
+            'imposto_renda' => ['required', 'numeric'],
+            'taxa_juros' => ['required', 'numeric'],
+            'taxa_deposito_compulsorio' => ['required', 'numeric'],
+            'salario_minimo' => ['required', 'numeric'],
         ];
     }
 }
